[Ref] Move RdfsResource from dto to model package.
