+++ src/main/java/cz/cvut/kbss/termit/rest/OidcUserController.java
@@ -19,7 +19,7 @@ package cz.cvut.kbss.termit.rest;
 
 
 import cz.cvut.kbss.jsonld.JsonLd;
-import cz.cvut.kbss.termit.dto.RdfsResource;
+import cz.cvut.kbss.termit.model.RdfsResource;
 import cz.cvut.kbss.termit.model.UserAccount;
 import cz.cvut.kbss.termit.security.SecurityConstants;
 import cz.cvut.kbss.termit.service.IdentifierResolver;
