+++ src/main/java/cz/cvut/kbss/termit/rest/VocabularyController.java
@@ -19,7 +19,7 @@ package cz.cvut.kbss.termit.rest;
 
 import cz.cvut.kbss.jsonld.JsonLd;
 import cz.cvut.kbss.termit.dto.AggregatedChangeInfo;
-import cz.cvut.kbss.termit.dto.RdfsStatement;
+import cz.cvut.kbss.termit.dto.RdfStatement;
 import cz.cvut.kbss.termit.dto.Snapshot;
 import cz.cvut.kbss.termit.dto.acl.AccessControlListDto;
 import cz.cvut.kbss.termit.dto.filter.ChangeRecordFilterDto;
@@ -459,10 +459,10 @@ public class VocabularyController extends BaseController {
             @ApiResponse(responseCode = "404", description = ApiDoc.ID_NOT_FOUND_DESCRIPTION),
     })
     @GetMapping(value = "/{localName}/relations")
-    public List<RdfsStatement> relations(@Parameter(description = ApiDoc.ID_LOCAL_NAME_DESCRIPTION,
+    public List<RdfStatement> relations(@Parameter(description = ApiDoc.ID_LOCAL_NAME_DESCRIPTION,
                                                     example = ApiDoc.ID_LOCAL_NAME_EXAMPLE)
                                          @PathVariable String localName,
-                                         @Parameter(description = ApiDoc.ID_NAMESPACE_DESCRIPTION,
+                                        @Parameter(description = ApiDoc.ID_NAMESPACE_DESCRIPTION,
                                                     example = ApiDoc.ID_NAMESPACE_EXAMPLE)
                                          @RequestParam(name = QueryParams.NAMESPACE,
                                                        required = false) Optional<String> namespace) {
@@ -479,10 +479,10 @@ public class VocabularyController extends BaseController {
             @ApiResponse(responseCode = "404", description = ApiDoc.ID_NOT_FOUND_DESCRIPTION),
     })
     @GetMapping(value = "/{localName}/terms/relations")
-    public List<RdfsStatement> termsRelations(@Parameter(description = ApiDoc.ID_LOCAL_NAME_DESCRIPTION,
+    public List<RdfStatement> termsRelations(@Parameter(description = ApiDoc.ID_LOCAL_NAME_DESCRIPTION,
                                                          example = ApiDoc.ID_LOCAL_NAME_EXAMPLE)
                                               @PathVariable String localName,
-                                              @Parameter(description = ApiDoc.ID_NAMESPACE_DESCRIPTION,
+                                             @Parameter(description = ApiDoc.ID_NAMESPACE_DESCRIPTION,
                                                          example = ApiDoc.ID_NAMESPACE_EXAMPLE)
                                               @RequestParam(name = QueryParams.NAMESPACE,
                                                             required = false) Optional<String> namespace) {
