+++ src/test/java/cz/cvut/kbss/termit/service/LanguageServiceTest.java
@@ -19,7 +19,7 @@ package cz.cvut.kbss.termit.service;
 
 import cz.cvut.kbss.jopa.model.MultilingualString;
 import cz.cvut.kbss.ontodriver.model.LangString;
-import cz.cvut.kbss.termit.dto.RdfsResource;
+import cz.cvut.kbss.termit.model.RdfsResource;
 import cz.cvut.kbss.termit.environment.Environment;
 import cz.cvut.kbss.termit.environment.Generator;
 import cz.cvut.kbss.termit.exception.InvalidLanguageConstantException;
